+++ R/f_wrap.R
@@ -6,10 +6,15 @@
 #' @param x A vector of text strings.
 #' @param width A positive integer giving the target column for wrapping lines
 #' in the output.
+#' @param sep A new line separator (defaults to \code{"\\n"}.
 #' @param exdent A non-negative integer specifying the indentation of subsequent
 #' lines in paragraphs.
 #' @param indent A non-negative integer giving the indentation of the first line
 #' in a paragraph.
+#' @param equal.lines logical.  If \code{TRUE} the number of lines for each
+#' element will be made the same by appending additional '\\n' to those below
+#' the max number of lines.  This is useful for legend spacing.
+
 #' @param \ldots Other arguments passed to \code{\link[base]{strwrap}}.
 #' @return Returns a string vector with wrapped new line characters.
 #' @rdname f_wrap
@@ -52,19 +57,34 @@
 #' )
 #'
 #' }
-f_wrap <- function (x, width = 15, exdent = 0, indent = 0, ...) {
+f_wrap <- function (x, width = 15, sep = '\n', exdent = 0, indent = 0,
+    equal.lines = FALSE, ...) {
 
     nas <- is.na(x)
-    out <- unlist(lapply(x, function(y) {
-        paste(
-            strwrap(y, width = width, exdent = exdent, indent = indent, ...),
-            collapse = "\n"
-        )
-    }))
+
+    if (isTRUE(equal.lines)) {
+        out <- lapply(x, function(y) {
+
+                strwrap(y, width = width, exdent = exdent, indent = indent, ...)
+
+        })
+
+        lens <- unlist(lapply(out, length))
+        ml <- max(lens)
+        out <- unlist(Map(function(i, x) paste(c(paste(x, collapse = sep), rep(sep, ml - i)), collapse = ''), lens, out))
+    } else {
+        out <- lapply(x, function(y) {
+
+                unlist(paste(strwrap(y, width = width, exdent = exdent, indent = indent, ...), collapse = sep))
+
+        })
+    }
+
     out[nas] <- NA
     out
 
 }
+
 
 
 #' @export